+++ src/Router/Configure/Handlers.php
@@ -5,8 +5,10 @@ declare(strict_types=1);
 namespace Gacela\Router\Configure;
 
 use Exception;
+use Gacela\Router\Exceptions\MethodNotAllowed405Exception;
 use Gacela\Router\Exceptions\NotFound404Exception;
 use Gacela\Router\Handlers\FallbackExceptionHandler;
+use Gacela\Router\Handlers\MethodNotAllowed405ExceptionHandler;
 use Gacela\Router\Handlers\NotFound404ExceptionHandler;
 use Throwable;
 
@@ -40,6 +42,7 @@ final class Handlers
     private function addBuiltInHandlers(): void
     {
         $this->handle(NotFound404Exception::class, NotFound404ExceptionHandler::class);
+        $this->handle(MethodNotAllowed405Exception::class, MethodNotAllowed405ExceptionHandler::class);
         $this->handle(Exception::class, FallbackExceptionHandler::class);
         $this->handle(Throwable::class, FallbackExceptionHandler::class);
     }